fix PERSIST return value for non-expiring key
